Move cpu streams executor to new API
